--- a/ggml/src/ggml-cuda/mmq_id.cu
+++ b/ggml/src/ggml-cuda/mmq_id.cu
@@ -201,6 +201,9 @@ static void ggml_cuda_mul_mat_q_switch_type_id(ggml_backend_cuda_context & ctx, 
         case GGML_TYPE_IQ2_KS:
             mul_mat_q_case_id<GGML_TYPE_IQ2_KS>(ctx, args, stream);
             break;
+        case GGML_TYPE_IQ2_KL:
+            mul_mat_q_case_id<GGML_TYPE_IQ2_KL>(ctx, args, stream);
+            break;
         case GGML_TYPE_IQ2_K:
             mul_mat_q_case_id<GGML_TYPE_IQ2_K>(ctx, args, stream);
             break;
@@ -433,6 +436,7 @@ bool ggml_cuda_can_use_mmq_id(enum ggml_type type, int cc, int64_t ne11) {
         case GGML_TYPE_IQ4_XS:
         case GGML_TYPE_IQ4_NL:
         case GGML_TYPE_IQ2_KS:
+        case GGML_TYPE_IQ2_KL:
         case GGML_TYPE_IQ2_K:
         case GGML_TYPE_IQ2_K_R4:
             mmq_supported = true;