--- a/ggml/src/ggml-cuda/mmq_id.cu
+++ b/ggml/src/ggml-cuda/mmq_id.cu
@@ -210,6 +210,9 @@ static void ggml_cuda_mul_mat_q_switch_type_id(ggml_backend_cuda_context & ctx, 
         case GGML_TYPE_IQ2_K_R4:
             mul_mat_q_case_id<GGML_TYPE_IQ2_K_R4>(ctx, args, stream);
             break;
+        case GGML_TYPE_IQ3_KS:
+            mul_mat_q_case_id<GGML_TYPE_IQ3_KS>(ctx, args, stream);
+            break;
         default:
             GGML_ABORT("fatal error");
             break;
@@ -439,6 +442,7 @@ bool ggml_cuda_can_use_mmq_id(enum ggml_type type, int cc, int64_t ne11) {
         case GGML_TYPE_IQ2_KL:
         case GGML_TYPE_IQ2_K:
         case GGML_TYPE_IQ2_K_R4:
+        case GGML_TYPE_IQ3_KS:
             mmq_supported = true;
             break;
         default: