--- a/ggml/src/ggml-cuda/mmq_id.cu
+++ b/ggml/src/ggml-cuda/mmq_id.cu
@@ -210,6 +210,12 @@ static void ggml_cuda_mul_mat_q_switch_type_id(ggml_backend_cuda_context & ctx, 
         case GGML_TYPE_IQ2_K_R4:
             mul_mat_q_case_id<GGML_TYPE_IQ2_K_R4>(ctx, args, stream);
             break;
+        case GGML_TYPE_IQ3_K:
+            mul_mat_q_case_id<GGML_TYPE_IQ3_K>(ctx, args, stream);
+            break;
+        case GGML_TYPE_IQ3_K_R4:
+            mul_mat_q_case_id<GGML_TYPE_IQ3_K_R4>(ctx, args, stream);
+            break;
         case GGML_TYPE_IQ3_KS:
             mul_mat_q_case_id<GGML_TYPE_IQ3_KS>(ctx, args, stream);
             break;
@@ -443,6 +449,8 @@ bool ggml_cuda_can_use_mmq_id(enum ggml_type type, int cc, int64_t ne11) {
         case GGML_TYPE_IQ2_K:
         case GGML_TYPE_IQ2_K_R4:
         case GGML_TYPE_IQ3_KS:
+        case GGML_TYPE_IQ3_K:
+        case GGML_TYPE_IQ3_K_R4:
             mmq_supported = true;
             break;
         default: