--- a/ggml/src/ggml-cuda/mmq_id.cu
+++ b/ggml/src/ggml-cuda/mmq_id.cu
@@ -261,6 +261,12 @@ static void ggml_cuda_mul_mat_q_switch_type_id(ggml_backend_cuda_context & ctx, 
         case GGML_TYPE_IQ2_KT:
             mul_mat_q_case_id<GGML_TYPE_IQ2_KT>(ctx, args, stream);
             break;
+        case GGML_TYPE_IQ3_KT:
+            mul_mat_q_case_id<GGML_TYPE_IQ3_KT>(ctx, args, stream);
+            break;
+        case GGML_TYPE_IQ4_KT:
+            mul_mat_q_case_id<GGML_TYPE_IQ4_KT>(ctx, args, stream);
+            break;
         default:
             GGML_ABORT("fatal error");
             break;
@@ -507,6 +513,8 @@ bool ggml_cuda_can_use_mmq_id(enum ggml_type type, int cc, int64_t ne11) {
         case GGML_TYPE_IQ6_K:
         case GGML_TYPE_IQ1_KT:
         case GGML_TYPE_IQ2_KT:
+        case GGML_TYPE_IQ3_KT:
+        case GGML_TYPE_IQ4_KT:
             mmq_supported = true;
             break;
         default: