--- a/ggml/src/ggml-cuda/template-instances/mmq-instance-iq1_s_id.cu
+++ b/ggml/src/ggml-cuda/template-instances/mmq-instance-iq1_s_id.cu
@@ -3,3 +3,4 @@
 #include "../mmq_id_common.cuh"
 
 DECL_MMQ_CASE(GGML_TYPE_IQ1_S);
+DECL_MMQ_CASE(GGML_TYPE_IQ1_S_R4);