`ik_llama.cpp` has excellent CUDA prompt processing (PP) performance for MoE models when using large batch and u-batch sizes (>= 2048). But for smaller batch sizes PP performance rapidly decreases with decreasing u-batch size. This is not desirable as larger u-batches result in larger CUDA compute buffers, which reduces the maximum context length one can use and/or the number of MoE layers one can offload to the GPU.

This PR remedies the situation by bringing massive performance improvement for small u-batch sizes.

For this PR I'm standing on the shoulder of giants. Actually, it is a single giant, Johannes Gaessler. The PR derived from his [PR 15525](https://github.com/ggml-org/llama.cpp/pull/15525)  in mainline `llama.cpp`. But, as it is always the case these days, it required quite a bit of adaptation. All additional `ik_llama.cpp` quants needed to be added as well to the new kernels, which of course required to get rid of the almighty assumption in mainline code that all quantized tensor data consist of blocks of a fixed size arranged contiguously in memory.

At this point it is a bit of an append-only programming style, but I'll clean up in subsequent PRs.

Here is a performance comparison between the main branch and this PR for DeepSeek-Lite quantized with `Q4_0` as a function of u-batch size for a fixed batch size is 4096 on Linux. GPU is RTX-4080. Flash attention is on, `mla = 3, fmoe = 1`.  

| model                | n_batch | n_ubatch |          test |      t/s (main)  |      t/s (PR)    |  Speedup |
| -------------------- | ------: | -------: | ------------: | ---------------: | ---------------: | -------: |
| deepseek2 16B Q4_0   |    4096 |      128 |        pp4096 |  1783.77 ± 54.27 |  3088.74 ± 51.34 |  1.732   |
| deepseek2 16B Q4_0   |    4096 |      256 |        pp4096 |  2934.71 ± 19.21 |  5222.70 ± 29.80 |  1.780   |
| deepseek2 16B Q4_0   |    4096 |      512 |        pp4096 |  4701.68 ± 44.76 |  7626.49 ± 62.89 |  1.622   |
| deepseek2 16B Q4_0   |    4096 |     1024 |        pp4096 |  6996.28 ± 64.63 |  9704.90 ± 94.28 |  1.387   |
| deepseek2 16B Q4_0   |    4096 |     2048 |        pp4096 |  9107.72 ± 37.01 | 10395.52 ± 101.25|  1.144   |
| deepseek2 16B Q4_0   |    4096 |     4096 |        pp4096 | 10256.00 ± 14.23 | 10265.86 ± 13.76 |  1.000   |

The PR massively reduces the number of kernel launches during prompt processing. Hence, I wouldn't be surprised if on Windows, where kernel launches are much more expensive, the performance impact is much higher than on Linux. 

Worth noting that the new MoE implementation becomes less efficient than the existing implementation somewhere around u-batch size of 2000 tokens. Because of that, `ik_llama.cpp` switches to the old MoE implementation for `u-batch > 2048`. I expect that the threshold where the old implementation is better will be model dependent (most likely dependent on ration between total and active experts), but haven't studied this in detail, so left for a future tuning.