@ikawrakow 2025-08-26
@DocShotgun Thanks! There was the `fp8` header include missing. Can you try now?