@usrlocalben 2025-08-26
> No real changes for hybrid inference. Something made PP fall when I went from batch 1024 to 512, previously it was over 100 on qwen-235b and now it's 66, but I don't think that it was this commit.

I _think_ something happened in the CUDA graphs PR, but I have not taken the time to measure commits yet and find it.