@Ph0rk0z 2025-08-26
I may also be mistaking qwen 512 perf for GLM perf. I tried rolling back to https://github.com/ikawrakow/ik_llama.cpp/commit/23fe18ce83237879dc5e55c444de560f0ed736d5 with no change. I'll try to roll back to before cuda graphs later today.