@ikawrakow 2025-08-26
Because @DocShotgun brought up a comparison with mainline `llama.cpp`: for the MoE models I can run fully offloaded to my 16 GB RTX-4080 GPU, `ik_llama.cpp` is still faster (but yes, Johannes has been able to close the gap quite a bit since April, when the `llama.cpp` performance improvement initiative kicked in). Here is what I get for 3 different models:

  | model                | n_ubatch |            test |    t/s (llama.cpp)   |  t/s (ik_llama.cpp)  |  Speedup |
| -------------------- | -------: | --------------: | -------------------: | -------------------: | -------: |
| deepseek2 16B Q4_0   |      256 |          pp4096 |       5379.33 ± 7.51 |    5192.36 ± 116.79  |  0.965   |
| deepseek2 16B Q4_0   |      512 |          pp4096 |       7268.16 ± 5.54 |    7458.49 ± 144.03  |  1.026   |
| deepseek2 16B Q4_0   |     1024 |          pp4096 |      8463.82 ± 15.63 |    9635.03 ± 149.52  |  1.138   |
| deepseek2 16B Q4_0   |     2048 |          pp4096 |      8372.37 ± 23.71 |    10363.95 ± 140.65 |  1.238   |
| deepseek2 16B Q4_0   |     4096 |          pp4096 |      7471.83 ± 10.94 |    10257.31 ± 48.72  |  1.373   |
| qwen3moe 30B.A3B IQ2_XXS |      256 |          pp4096 |      3758.21 ± 67.68 |  3990.02 ± 18.09 |  1.061   |
| qwen3moe 30B.A3B IQ2_XXS |      512 |          pp4096 |     5167.30 ± 212.46 |  5576.11 ± 15.25 |  1.079   |
| qwen3moe 30B.A3B IQ2_XXS |     1024 |          pp4096 |     6039.60 ± 231.07 |   6518.85 ± 8.76 |  1.079   |
| qwen3moe 30B.A3B IQ2_XXS |     2048 |          pp4096 |      6299.04 ± 24.53 |  6587.33 ± 13.69 |  1.046   |
| qwen3moe 30B.A3B IQ2_XXS |     4096 |          pp4096 |       5846.55 ± 5.21 | 6082.93 ± 190.64 |  1.040   |
| gpt-oss 20B MXFP4 MoE    |      256 |          pp3072 |      4881.85 ± 23.06 |  5474.21 ± 35.55 |  1.119   |
| gpt-oss 20B MXFP4 MoE    |      512 |          pp3072 |       6573.29 ± 8.64 |  7282.90 ± 26.87 |  1.108   |
| gpt-oss 20B MXFP4 MoE    |     1024 |          pp3072 |      7512.86 ± 20.92 |  8266.53 ± 31.21 |  1.100   |
| gpt-oss 20B MXFP4 MoE    |     1536 |          pp3072 |      7562.65 ± 23.94 |  8878.06 ± 42.00 |  1.174   |
| gpt-oss 20B MXFP4 MoE    |     3072 |          pp3072 |      7200.01 ± 11.84 |  9364.51 ± 29.43 |  1.301   |