@DocShotgun 2025-08-26
> Because @DocShotgun brought up a comparison with mainline `llama.cpp`: for the MoE models I can run fully offloaded to my 16 GB RTX-4080 GPU, `ik_llama.cpp` is still faster (but yes, Johannes has been able to close the gap quite a bit since April, when the `llama.cpp` performance improvement initiative kicked in). Here is what I get for 3 different models:

Do you think mainline `llama.cpp` still has some other Windows-specific optimization that hasn't been ported over yet?